alert-manager/requirements.txt
redis==5.0.1
pyyaml==6.0.1
requests==2.31.0
jinja2==3.1.2
structlog==23.2.0
prometheus-client==0.19.0
python-dotenv==1.0.0

# Email
aiosmtplib==3.0.1

# Testing
pytest==7.4.3
pytest-mock==3.12.0
